re-arranging structures and variables
